gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aio-pika==9.4.1
httpx==0.27.0
PyJWT==2.8.0
